moved render_plot functionality and tests into dataset_renderer
