+++ tests/integration/visualise/test_render_map.py
@@ -6,3 +6,5 @@ Integration tests for test_render_map.py
 # as we need a server to be able to see the images.  As that's the case,
 # integration tests for this module will be put on hold until we have a server
 # to render images on and test.
+
+
